+++ mkbak/__init__.py
@@ -21,4 +21,4 @@ copied = []
 deleted = []
 errors = []
 warnings = []
-VERSION = "1.2.3"
+VERSION = "1.3.0"